Move validation tests to BaseCreateSerialPublicationHandlerTest
