+++ src/test/java/no/sikt/nva/pubchannels/handler/create/journal/CreateJournalHandlerTest.java
@@ -69,103 +69,6 @@ class CreateJournalHandlerTest extends BaseCreateSerialPublicationHandlerTest {
         assertThat(response.getBodyObject(SerialPublicationDto.class), is(equalTo(expectedJournal)));
     }
 
-    @ParameterizedTest(name = "Should return BadGateway for response code \"{0}\"")
-    @ValueSource(ints = {HttpURLConnection.HTTP_UNAUTHORIZED, HttpURLConnection.HTTP_INTERNAL_ERROR,
-        HttpURLConnection.HTTP_UNAVAILABLE})
-    void shouldReturnBadGatewayWhenAuthResponseNotSuccessful(int httpStatusCode) throws IOException {
-        var input = constructRequest(new CreateSerialPublicationRequestBuilder().withName(VALID_NAME).build());
-
-        stubAuth(httpStatusCode);
-
-        handlerUnderTest.handleRequest(input, output, context);
-
-        var response = GatewayResponse.fromOutputStream(output, Problem.class);
-
-        var statusCode = response.getStatusCode();
-        assertThat(statusCode, is(equalTo(HttpURLConnection.HTTP_BAD_GATEWAY)));
-
-        var problem = response.getBodyObject(Problem.class);
-
-        assertThat(problem.getDetail(), is(equalTo("Unexpected response from upstream!")));
-    }
-
-    @Test
-    void shouldReturnBadGatewayWhenAuthClientInterruptionOccurs() throws IOException, InterruptedException {
-        this.handlerUnderTest = new CreateJournalHandler(environment, setupInteruptedClient());
-
-        var input = constructRequest(new CreateSerialPublicationRequestBuilder().withName(VALID_NAME).build());
-
-        var appender = LogUtils.getTestingAppenderForRootLogger();
-
-        handlerUnderTest.handleRequest(input, output, context);
-
-        assertThat(appender.getMessages(), containsString("Unable to reach upstream!"));
-        assertThat(appender.getMessages(), containsString(InterruptedException.class.getSimpleName()));
-
-        var response = GatewayResponse.fromOutputStream(output, Problem.class);
-
-        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_BAD_GATEWAY)));
-
-        var problem = response.getBodyObject(Problem.class);
-        assertThat(problem.getDetail(), is(equalTo("Unable to reach upstream!")));
-    }
-
-    @ParameterizedTest(name = "Should return BadRequest for invalid name \"{0}\"")
-    @MethodSource("invalidNames")
-    void shouldReturnBadRequestWhenNameInvalid(String name) throws IOException {
-
-        var requestBody = new CreateSerialPublicationRequestBuilder().withName(name).build();
-        handlerUnderTest.handleRequest(constructRequest(requestBody), output, context);
-        var response = GatewayResponse.fromOutputStream(output, Problem.class);
-
-        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_BAD_REQUEST)));
-
-        var problem = response.getBodyObject(Problem.class);
-        assertThat(problem.getDetail(), is(containsString("Name is too")));
-    }
-
-    @ParameterizedTest(name = "Should return BadRequest for invalid print ISSN \"{0}\"")
-    @MethodSource("invalidIssn")
-    void shouldReturnBadRequestWhenInvalidPissn(String issn) throws IOException {
-        var requestBody =
-            new CreateSerialPublicationRequestBuilder().withName(VALID_NAME).withPrintIssn(issn).build();
-        handlerUnderTest.handleRequest(constructRequest(requestBody), output, context);
-        var response = GatewayResponse.fromOutputStream(output, Problem.class);
-
-        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_BAD_REQUEST)));
-
-        var problem = response.getBodyObject(Problem.class);
-        assertThat(problem.getDetail(), is(containsString("PrintIssn has an invalid ISSN format")));
-    }
-
-    @ParameterizedTest(name = "Should return BadRequest for invalid online ISSN \"{0}\"")
-    @MethodSource("invalidIssn")
-    void shouldReturnBadRequestWhenInvalidElectronicIssn(String issn) throws IOException {
-        var requestBody =
-            new CreateSerialPublicationRequestBuilder().withName(VALID_NAME).withOnlineIssn(issn).build();
-        handlerUnderTest.handleRequest(constructRequest(requestBody), output, context);
-        var response = GatewayResponse.fromOutputStream(output, Problem.class);
-
-        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_BAD_REQUEST)));
-
-        var problem = response.getBodyObject(Problem.class);
-        assertThat(problem.getDetail(), is(containsString("OnlineIssn has an invalid ISSN format")));
-    }
-
-    @ParameterizedTest(name = "Should return BadRequest for invalid URL \"{0}\"")
-    @MethodSource("invalidUri")
-    void shouldReturnBadRequestWhenInvalidUrl(String url) throws IOException {
-        var requestBody =
-            new CreateSerialPublicationRequestBuilder().withName(VALID_NAME).withHomepage(url).build();
-        handlerUnderTest.handleRequest(constructRequest(requestBody), output, context);
-        var response = GatewayResponse.fromOutputStream(output, Problem.class);
-
-        assertThat(response.getStatusCode(), is(equalTo(HttpURLConnection.HTTP_BAD_REQUEST)));
-
-        var problem = response.getBodyObject(Problem.class);
-        assertThat(problem.getDetail(), is(containsString("Homepage has an invalid URL format")));
-    }
-
     @ParameterizedTest(name = "Should create series for print ISSN \"{0}\"")
     @MethodSource("validIssn")
     void shouldCreateJournalWithNameAndPrintIssn(String issn) throws IOException {
